maj policy

+++ Atelier17_Laravel/resources/views/articles/voirplus.blade.php
@@ -26,7 +26,7 @@
                     <div
                         class="card-header p-0 position-relative mt-n4 mx-3 z-index-2 bg-transparent
                     d-flex justify-content-center align-items-center">
-                        <img src="{{ asset('images/' . $article->image) }}" style="width: 400px" alt="image de l'article">
+                        <img src="{{ asset('images/' . $article->image) }}" style="width: 400px;max-width: 100%; max-height: 100%;" alt="image de l'article">
                     </div>
                     <div class="card-body">
                         <h6 class="mb-0 ">{{ $article->nom }}</h6>
@@ -43,7 +43,7 @@
                             <a href="{{ '/article/modifier/' . $article->id }}" class="btn btn-primary me-3">modifier</a>
                             <a href="/articles/deletearticle/{{ $article->id }}" class="btn btn-primary">Supprimer</a>
             @if (isset($ok) && $commentaire->user_id === Auth::user()->id)
-                <form action="{{ url('/articles/commentaireupdate/' . $commentaire->id) }}" class="comment_class" method="post">
+                <form action="{{ url('/articles/commentaireupdate/' . $commentaire->id) }}" class="comment_class" style="display:flex;justify-content:space-around;flex-wrap:nowrap;" method="post">
                     @csrf
                     <input type="hidden" name="article_id" value="{{ $article->id }}">
                     <input type="hidden" name="id" value='{{ $commentaire->id }}'>
@@ -59,8 +59,10 @@
             <div class="comment" style="display:flex;justify-content:space-around;flex-wrap:nowrap;">
                 <div class="mb-1">Contenu : {{ $comment->contenu }}</div>
                 <div class="text-muted mb-1">Auteur : {{ $comment->user->name }}</div>
-                <div class="text-muted mb-1"> Date de publication : {{ $comment->created_at }}</div> 
+                <div class="text-muted mb-1"> Date de publication : {{ $comment->created_at }}</div>
+                @can('delete', $comment)
                     <a href="/articles/deletecommentaire/{{ $comment->id }}" class="badge rounded-pill bg-primary">Supprimer</a>
+                @endcan
                     <a href="/articles/commentaire/{{$comment->id}}" class="badge rounded-pill bg-dark">Modifier</a>
             </div>
         @empty
